+++ bindings/nodejs/examples/how_tos/nfts/burn_nft.ts
@@ -5,7 +5,7 @@ import { Wallet } from '@iota/sdk';
 
 // In this example we will burn an existing nft output.
 //
-// Make sure that `example.stronghold` and `example.walletdb` already exist by
+// Make sure that `STRONGHOLD_SNAPSHOT_PATH` and `WALLET_DB_PATH` already exist by
 // running the `how_tos/accounts-and-addresses/create-wallet` example!
 //
 // Rename `.env.example` to `.env` first, then run