+++ bindings/python/examples/how_tos/client/get_health.py
@@ -1,6 +1,5 @@
 from iota_sdk import Client
 from dotenv import load_dotenv
-import json
 import os
 
 load_dotenv()