+++ bindings/python/iota_sdk/__init__.py
@@ -7,9 +7,11 @@ from .client._high_level_api import GenerateAddressesOptions, GenerateAddressOpt
 from .utils import Utils
 from .wallet.wallet import Wallet, Account
 from .wallet.common import WalletError
+from .wallet.sync_options import AccountSyncOptions, AliasSyncOptions, NftSyncOptions, SyncOptions
 from .secret_manager.secret_manager import *
 from .prefix_hex import *
 from .types.address import *
+from .types.block import *
 from .types.burn import *
 from .types.common import *
 from .types.feature import *