+++ sdk/examples/client/high_level/search_address.rs
@@ -3,12 +3,15 @@
 
 //! In this example we will try to find the index and address type of an address.
 //!
-//! `cargo run --example search_address --release`
+//! Rename `.env.example` to `.env` first, then run the command:
+//! ```sh
+//! cargo run --release --example search_address [BECH32_ADDRESS] [START_INDEX] [RANGE_SIZE]
+//! ```
 
 use iota_sdk::client::{
-    api::search_address,
-    constants::IOTA_COIN_TYPE,
-    secret::{mnemonic::MnemonicSecretManager, SecretManager},
+    api::{search_address, GetAddressesOptions},
+    constants::SHIMMER_COIN_TYPE,
+    secret::SecretManager,
     Client, Result,
 };
 
@@ -19,33 +22,38 @@ async fn main() -> Result<()> {
 
     let node_url = std::env::var("NODE_URL").unwrap();
 
-    // Create a client instance
-    let client = Client::builder()
-        .with_node(&node_url)? // Insert your node URL here
-        .finish()?;
+    // Create a node client.
+    let client = Client::builder().with_node(&node_url)?.finish().await?;
 
     let secret_manager =
         SecretManager::try_from_mnemonic(std::env::var("NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1").unwrap())?;
 
-    let addresses = client
-        .get_addresses(&secret_manager)
-        .with_account_index(0)
-        .with_range(9..10)
-        .get_raw()
-        .await?;
+    let address = if let Some(addr) = std::env::args()
+        .nth(1)
+        .map(|addr| addr.parse().expect("invalid address"))
+    {
+        addr
+    } else {
+        secret_manager
+            .generate_ed25519_addresses(GetAddressesOptions::from_client(&client).await?.with_range(0..1))
+            .await?[0]
+    };
+    println!("Search address: {address:#?}");
 
-    println!("{:?}", addresses[0]);
+    let address_range_start = std::env::args().nth(2).map(|s| s.parse::<u32>().unwrap()).unwrap_or(0);
+    let address_range_len = std::env::args().nth(3).map(|s| s.parse::<u32>().unwrap()).unwrap_or(10);
+    let address_range = address_range_start..address_range_start + address_range_len;
 
-    let res = search_address(
+    let (address_index, is_internal) = search_address(
         &secret_manager,
-        &client.get_bech32_hrp().await?,
-        IOTA_COIN_TYPE,
+        client.get_bech32_hrp().await?,
+        SHIMMER_COIN_TYPE,
         0,
-        0..10,
-        &addresses[0],
+        address_range,
+        &address.into(),
     )
     .await?;
+    println!("Address index: {address_index}\nIs internal address: {is_internal}");
 
-    println!("Address index: {}\nIs internal address: {}", res.0, res.1);
     Ok(())
 }