+++ sdk/examples/how_tos/advanced_transactions/advanced_transaction.rs
@@ -6,8 +6,6 @@
 //!
 //! `cargo run --release --all-features --example advanced_transaction`
 
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
 use iota_sdk::{
     types::block::{
         address::Bech32Address,
@@ -40,8 +38,8 @@ async fn main() -> Result<()> {
             .await?;
 
         // Create an ouput with amount 1_000_000 and a timelock of 1 hour
-        let in_an_hour = (SystemTime::now() + Duration::from_secs(3600))
-            .duration_since(UNIX_EPOCH)
+        let in_an_hour = (std::time::SystemTime::now() + std::time::Duration::from_secs(3600))
+            .duration_since(std::time::UNIX_EPOCH)
             .expect("clock went backwards")
             .as_secs()
             .try_into()