+++ sdk/examples/wallet/17_check_unlock_conditions.rs
@@ -3,14 +3,12 @@
 
 //! In this example we check if an output has only an address unlock condition and that the address is from the account.
 //!
-//! Make sure that `example.stronghold` and `example.walletdb` already exist by
-//! running the `create_account` example!
+//! Make sure that `STRONGHOLD_SNAPSHOT_PATH` and `WALLET_DB_PATH` already exist by
+//! running the `./how_tos/accounts_and_addresses/create_account.rs` example!
 //!
 //! ```sh
 //! cargo run --release --all-features --example check_unlock_conditions
 //! ```
-
-use std::env::var;
 
 use iota_sdk::{
     types::block::{
@@ -30,7 +28,7 @@ async fn main() -> Result<()> {
     dotenvy::dotenv().ok();
 
     let wallet = Wallet::builder()
-        .with_storage_path(&var("WALLET_DB_PATH").unwrap())
+        .with_storage_path(&std::env::var("WALLET_DB_PATH").unwrap())
         .finish()
         .await?;
     let account = wallet.get_account("Alice").await?;