+++ sdk/examples/wallet/participation.rs
@@ -7,8 +7,8 @@
 //! * try to vote (the example aborts if you vote on an already ended voting)
 //! * if a voting occurred, stops the voting and destroys the voting output
 //!
-//! Make sure that `example.stronghold` and `example.walletdb` already exist by
-//! running the `create_account` example and there are funds on the first address
+//! Make sure that `STRONGHOLD_SNAPSHOT_PATH` and `WALLET_DB_PATH` already exist by
+//! running the `./how_tos/accounts_and_addresses/create_account.rs` example and there are funds on the first address
 //! by running the `get_funds` example!
 //!
 //! Rename `.env.example` to `.env` first, then run the command:
@@ -16,11 +16,8 @@
 //! cargo run --release --all-features --example wallet_participation
 //! ```
 
-use std::{env::var, str::FromStr};
-
 use iota_sdk::{
     client::node_manager::node::Node,
-    types::api::plugins::participation::types::ParticipationEventId,
     wallet::{account::types::participation::ParticipationEventRegistrationOptions, Result},
     Url, Wallet,
 };
@@ -45,17 +42,17 @@ async fn main() -> Result<()> {
     dotenvy::dotenv().ok();
 
     let wallet = Wallet::builder()
-        .with_storage_path(&var("WALLET_DB_PATH").unwrap())
+        .with_storage_path(&std::env::var("WALLET_DB_PATH").unwrap())
         .finish()
         .await?;
     let account = wallet.get_account("Alice").await?;
 
     // Provide the stronghold password
     wallet
-        .set_stronghold_password(var("STRONGHOLD_PASSWORD").unwrap())
+        .set_stronghold_password(std::env::var("STRONGHOLD_PASSWORD").unwrap())
         .await?;
 
-    let event_id = ParticipationEventId::from_str(PARTICIPATION_EVENT_ID_1)?;
+    let event_id = PARTICIPATION_EVENT_ID_1.parse()?;
     let node = Node {
         url: Url::parse(PARTICPATION_NODE_URL).map_err(iota_sdk::client::Error::Url)?,
         auth: None,
@@ -66,7 +63,7 @@ async fn main() -> Result<()> {
             node,
             // We ignore this particular event
             events_to_ignore: (!IGNORED_PARTICIPATION_EVENT_ID.is_empty())
-                .then_some(vec![ParticipationEventId::from_str(IGNORED_PARTICIPATION_EVENT_ID)?]),
+                .then_some(vec![IGNORED_PARTICIPATION_EVENT_ID.parse()?]),
             // We register all others. If you want to register only particular events provide their ids with a
             // `Some(vec![...])`
             events_to_register: None,
@@ -103,7 +100,7 @@ async fn main() -> Result<()> {
     ////////////////////////////////////////////////
     if !DEREGISTERED_PARTICIPATION_EVENT.is_empty() {
         account
-            .deregister_participation_event(&ParticipationEventId::from_str(DEREGISTERED_PARTICIPATION_EVENT)?)
+            .deregister_participation_event(&DEREGISTERED_PARTICIPATION_EVENT.parse()?)
             .await?;
 
         println!("Registered events (updated):");
@@ -139,8 +136,8 @@ async fn main() -> Result<()> {
         .retry_transaction_until_included(&transaction.transaction_id, None, None)
         .await?;
     println!(
-        "Transaction included: {}/block/{}",
-        var("EXPLORER_URL").unwrap(),
+        "Block included: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
         block_id
     );
 
@@ -164,8 +161,8 @@ async fn main() -> Result<()> {
         .retry_transaction_until_included(&transaction.transaction_id, None, None)
         .await?;
     println!(
-        "Transaction included: {}/block/{}",
-        var("EXPLORER_URL").unwrap(),
+        "Block included: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
         block_id
     );
 
@@ -189,8 +186,8 @@ async fn main() -> Result<()> {
         .retry_transaction_until_included(&transaction.transaction_id, None, None)
         .await?;
     println!(
-        "Transaction included: {}/block/{}",
-        var("EXPLORER_URL").unwrap(),
+        "Block included: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
         block_id
     );
 
@@ -216,8 +213,8 @@ async fn main() -> Result<()> {
         .retry_transaction_until_included(&transaction.transaction_id, None, None)
         .await?;
     println!(
-        "Transaction included: {}/block/{}",
-        var("EXPLORER_URL").unwrap(),
+        "Block included: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
         block_id
     );
 
@@ -240,8 +237,8 @@ async fn main() -> Result<()> {
         .retry_transaction_until_included(&transaction.transaction_id, None, None)
         .await?;
     println!(
-        "Transaction included: {}/block/{}",
-        var("EXPLORER_URL").unwrap(),
+        "Block included: {}/block/{}",
+        std::env::var("EXPLORER_URL").unwrap(),
         block_id
     );